Split test file
