+++ tests/open_alchemy/models_file/model/test_source/test_source.py
@@ -18,362 +18,6 @@ _ModelArtifacts = models_file.types.ModelArtifacts
     "artifacts, expected_source",
     [
         (
-            _TypedDictArtifacts(
-                required=_TypedDictClassArtifacts(
-                    props=[_ColumnArtifacts(name="column_1", type="type_1")],
-                    empty=False,
-                    name="ModelRequiredDict",
-                    parent_class="RequiredParentClass",
-                ),
-                not_required=None,  # type: ignore
-            ),
-            '''
-
-class ModelRequiredDict(RequiredParentClass, total=True):
-    """TypedDict for properties that are required."""
-
-    column_1: type_1''',
-        ),
-        (
-            _TypedDictArtifacts(
-                required=_TypedDictClassArtifacts(
-                    props=[
-                        _ColumnArtifacts(name="column_1", type="type_1"),
-                        _ColumnArtifacts(name="column_2", type="type_2"),
-                    ],
-                    empty=False,
-                    name="ModelRequiredDict",
-                    parent_class="RequiredParentClass",
-                ),
-                not_required=None,  # type: ignore
-            ),
-            '''
-
-class ModelRequiredDict(RequiredParentClass, total=True):
-    """TypedDict for properties that are required."""
-
-    column_1: type_1
-    column_2: type_2''',
-        ),
-    ],
-    ids=["single property", "multiple properties"],
-)
-@pytest.mark.models_file
-def test_typed_dict_required(artifacts, expected_source):
-    """
-    GIVEN model artifacts
-    WHEN typed_dict_required is called with the artifacts
-    THEN the source code for the typed dict class is returned.
-    """
-    source = models_file._model._source.typed_dict_required(artifacts=artifacts)
-
-    assert source == expected_source
-
-
-@pytest.mark.parametrize(
-    "artifacts, expected_source",
-    [
-        (
-            _TypedDictArtifacts(
-                required=None,  # type: ignore
-                not_required=_TypedDictClassArtifacts(
-                    props=[],
-                    empty=True,
-                    name="ModelNotRequiredDict",
-                    parent_class="NotRequiredParentClass",
-                ),
-            ),
-            '''
-
-class ModelNotRequiredDict(NotRequiredParentClass, total=False):
-    """TypedDict for properties that are not required."""''',
-        ),
-        (
-            _TypedDictArtifacts(
-                required=None,  # type: ignore
-                not_required=_TypedDictClassArtifacts(
-                    props=[_ColumnArtifacts(name="column_1", type="type_1")],
-                    empty=False,
-                    name="ModelNotRequiredDict",
-                    parent_class="NotRequiredParentClass",
-                ),
-            ),
-            '''
-
-class ModelNotRequiredDict(NotRequiredParentClass, total=False):
-    """TypedDict for properties that are not required."""
-
-    column_1: type_1''',
-        ),
-        (
-            _TypedDictArtifacts(
-                required=None,  # type: ignore
-                not_required=_TypedDictClassArtifacts(
-                    props=[
-                        _ColumnArtifacts(name="column_1", type="type_1"),
-                        _ColumnArtifacts(name="column_2", type="type_2"),
-                    ],
-                    empty=False,
-                    name="ModelNotRequiredDict",
-                    parent_class="NotRequiredParentClass",
-                ),
-            ),
-            '''
-
-class ModelNotRequiredDict(NotRequiredParentClass, total=False):
-    """TypedDict for properties that are not required."""
-
-    column_1: type_1
-    column_2: type_2''',
-        ),
-    ],
-    ids=["empty", "single property", "multiple properties"],
-)
-@pytest.mark.models_file
-def test_typed_dict_not_required(artifacts, expected_source):
-    """
-    GIVEN model artifacts
-    WHEN typed_dict_not_required is called with the artifacts
-    THEN the source code for the typed dict class is returned.
-    """
-    source = models_file._model._source.typed_dict_not_required(artifacts=artifacts)
-
-    assert source == expected_source
-
-
-@pytest.mark.parametrize(
-    "artifacts, expected_source",
-    [
-        (_ArgArtifacts(required=[], not_required=[]), ""),
-        (
-            _ArgArtifacts(
-                required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    )
-                ],
-                not_required=[],
-            ),
-            ", column_1: init_type_1",
-        ),
-        (
-            _ArgArtifacts(
-                required=[],
-                not_required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    )
-                ],
-            ),
-            ", column_1: init_type_1 = None",
-        ),
-        (
-            _ArgArtifacts(
-                required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    ),
-                    _ColumnArgArtifacts(
-                        name="column_2",
-                        init_type="init_type_2",
-                        from_dict_type="fd_type_2",
-                    ),
-                ],
-                not_required=[],
-            ),
-            ", column_1: init_type_1, column_2: init_type_2",
-        ),
-        (
-            _ArgArtifacts(
-                required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    )
-                ],
-                not_required=[
-                    _ColumnArgArtifacts(
-                        name="column_2",
-                        init_type="init_type_2",
-                        from_dict_type="fd_type_2",
-                    )
-                ],
-            ),
-            ", column_1: init_type_1, column_2: init_type_2 = None",
-        ),
-        (
-            _ArgArtifacts(
-                required=[],
-                not_required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    ),
-                    _ColumnArgArtifacts(
-                        name="column_2",
-                        init_type="init_type_2",
-                        from_dict_type="fd_type_2",
-                    ),
-                ],
-            ),
-            ", column_1: init_type_1 = None, column_2: init_type_2 = None",
-        ),
-    ],
-    ids=[
-        "empty",
-        "single required",
-        "single not required",
-        "multiple required",
-        "multiple required and not required",
-        "multiple not required",
-    ],
-)
-@pytest.mark.models_file
-def test_arg_init(artifacts, expected_source):
-    """
-    GIVEN artifacts
-    WHEN arg_init is called with the artifacts
-    THEN the expected source is returned.
-    """
-    source = models_file._model._source.arg_init(artifacts=artifacts)
-
-    assert source == expected_source
-
-
-@pytest.mark.parametrize(
-    "artifacts, expected_source",
-    [
-        (_ArgArtifacts(required=[], not_required=[]), ""),
-        (
-            _ArgArtifacts(
-                required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    )
-                ],
-                not_required=[],
-            ),
-            ", column_1: fd_type_1",
-        ),
-        (
-            _ArgArtifacts(
-                required=[],
-                not_required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    )
-                ],
-            ),
-            ", column_1: fd_type_1 = None",
-        ),
-        (
-            _ArgArtifacts(
-                required=[],
-                not_required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                        default='"value 1"',
-                    )
-                ],
-            ),
-            ', column_1: fd_type_1 = "value 1"',
-        ),
-        (
-            _ArgArtifacts(
-                required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    ),
-                    _ColumnArgArtifacts(
-                        name="column_2",
-                        init_type="init_type_2",
-                        from_dict_type="fd_type_2",
-                    ),
-                ],
-                not_required=[],
-            ),
-            ", column_1: fd_type_1, column_2: fd_type_2",
-        ),
-        (
-            _ArgArtifacts(
-                required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    )
-                ],
-                not_required=[
-                    _ColumnArgArtifacts(
-                        name="column_2",
-                        init_type="init_type_2",
-                        from_dict_type="fd_type_2",
-                    )
-                ],
-            ),
-            ", column_1: fd_type_1, column_2: fd_type_2 = None",
-        ),
-        (
-            _ArgArtifacts(
-                required=[],
-                not_required=[
-                    _ColumnArgArtifacts(
-                        name="column_1",
-                        init_type="init_type_1",
-                        from_dict_type="fd_type_1",
-                    ),
-                    _ColumnArgArtifacts(
-                        name="column_2",
-                        init_type="init_type_2",
-                        from_dict_type="fd_type_2",
-                    ),
-                ],
-            ),
-            ", column_1: fd_type_1 = None, column_2: fd_type_2 = None",
-        ),
-    ],
-    ids=[
-        "empty",
-        "single required",
-        "single not required",
-        "single default",
-        "multiple required",
-        "multiple required and not required",
-        "multiple not required",
-    ],
-)
-@pytest.mark.models_file
-def test_arg_from_dict(artifacts, expected_source):
-    """
-    GIVEN artifacts
-    WHEN arg_from_dict is called with the artifacts
-    THEN the expected source is returned.
-    """
-    source = models_file._model._source.arg_from_dict(artifacts=artifacts)
-
-    assert source == expected_source
-
-
-@pytest.mark.parametrize(
-    "artifacts, expected_source",
-    [
-        (
             _ModelArtifacts(
                 sqlalchemy=_SQLAlchemyModelArtifacts(
                     name="Model",
